+++ src/app/api.js
@@ -209,7 +209,9 @@ export const configs = {
         props: {
           transpose: true,
         },
-        x: 2, y: 4, w: 10, h: 2 },
+        x: 2, y: 4, w: 5, h: 2 },
+      { component: 'expressionHistogram',
+        x: 7, y: 4, w: 5, h: 2 },
       { component: 'scatterplot',
         coordinationScopes: {
           embeddingType: 'PCA',
@@ -500,7 +502,9 @@ export const configs = {
         },
         x: 0, y: 0, w: 10, h: 2 },
       { component: 'genes',
-        x: 10, y: 0, w: 2, h: 2 },
+        x: 10, y: 0, w: 2, h: 4 },
+      { component: 'expressionHistogram',
+        x: 0, y: 2, w: 10, h: 2 },
     ],
   },